refactor: extract EpiDataCall into _call.py
Verbatim move of EpiDataCall and _request_with_retry out of request.py
into a new _call.py module; request.py imports them back. Breaks the
EpiDataCall<->endpoints import cycle so the endpoint class can construct
calls directly in the next commit.
